adjust authentication methods so that only owners have Update and delete as well as osme admin auth checks

+++ CheckAhead/src/main/java/com/skilldistillery/checkahead/services/RatingServiceImpl.java
@@ -21,14 +21,14 @@ public class RatingServiceImpl implements RatingService {
 
 	@Override
 	public Rating updateRating(Rating rating, int id, String username) {
-		Optional<Rating> oldRating = rateRepo.findById(id);
-		Rating newRating = null;
-		if(oldRating.isPresent()) {
-			newRating = oldRating.get();
-			newRating.setId(id);
-			newRating.setCategory(rating.getCategory());
-			if (userRepo.findByUsername(username).getUsername().equals(username)) {
-				return rateRepo.saveAndFlush(rating);
+		if (userRepo.findByUsername(username).getRole().equals("admin")) {
+			Optional<Rating> oldRating = rateRepo.findById(id);
+			Rating newRating = null;
+			if(oldRating.isPresent()) {
+				newRating = oldRating.get();
+				newRating.setId(id);
+				newRating.setCategory(rating.getCategory());
+				return rateRepo.saveAndFlush(newRating);
 			}
 		}
 		return null;
@@ -37,12 +37,13 @@ public class RatingServiceImpl implements RatingService {
 	@Override
 	public boolean deleteRating(int id, String username) {
 		boolean answer = false;
-		Optional<Rating> rating = rateRepo.findById(id);
-		if (rating.isPresent() && userRepo.findByUsername(username).getUsername().equals(username)) {
-			rateRepo.deleteById(id);
-			answer = true;
+		if (userRepo.findByUsername(username).getRole().equals("admin")) {
+			Optional<Rating> rating = rateRepo.findById(id);
+			if (rating.isPresent() && userRepo.findByUsername(username).getUsername().equals(username)) {
+				rateRepo.deleteById(id);
+				answer = true;
+			} 
 		}
-
 		return answer;
 
 	}
@@ -53,7 +54,6 @@ public class RatingServiceImpl implements RatingService {
 		if (rating.isPresent()) {
 			return rating.get();
 		} else {
-
 			return null;
 		}
 	}
@@ -61,11 +61,12 @@ public class RatingServiceImpl implements RatingService {
 	@Override
 	public Rating createRating(Rating rating, String username) {
 		Rating newRating = null;
-		if (userRepo.findByUsername(username).getUsername().equals(username)) {
-			newRating = rateRepo.saveAndFlush(rating);
+		if (userRepo.findByUsername(username).getRole().equals("admin")) {
+			if (userRepo.findByUsername(username).getUsername().equals(username)) {
+				newRating = rateRepo.saveAndFlush(rating);
+			} 
 		}
 		return newRating;
-
 	}
